+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 - `Gradient.isAGradient` checks if a value can be used as a Gradient
 - `Gradient.getMaxColorSequenceKeypoints` returns the maximum number of keypoints a ColorSequence can have
 - Added the [HSLuv and HPLuv](https://www.hsluv.org/) color spaces
+- Added the [Oklab](https://bottosson.github.io/posts/oklab/) color space
 
 ### Changed
 - Code refactoring and module consolidation